third_party/blink/web_tests/virtual/controls-refresh/calendar-picker/date-picker-open-with-enter-key.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>

<input type=date id=date value="2019-09-10">

<script>

test(() => {
  document.getElementById('date').focus();
  eventSender.keyDown("Enter");
  if (internals.runtimeFlags.formControlsRefreshEnabled) {
    assert_not_equals(internals.pagePopupWindow, null);
  } else {
    assert_equals(internals.pagePopupWindow, null);
  }
}, "Test opening date picker with Enter key.");

</script>
</body>
</html>
